Moved the attachment list cell to userinterface
Attachment list cell deals with user interface so I moved it there.
